src/backend/api/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.34.0
psycopg2-binary>=2.9.9
pydantic>=2.0.0
pydantic-settings>=2.0.0
